mockgun: percent are int
From the doc of the api:
```
percent
value:	int | None
range:	-2147483648, 2147483647
```

+++ shotgun_api3/lib/mockgun/mockgun.py
@@ -460,7 +460,7 @@ class Shotgun(object):
                     python_type = {"number": int,
                                    "float": float,
                                    "checkbox": bool,
-                                   "percent": float,
+                                   "percent": int,
                                    "text": basestring,
                                    "serializable": dict,
                                    "date": datetime.date,
